start working on reservation calendar
