+++ client/src/components/App.jsx
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import DatePickerForm from './DatePickerForm.jsx'
+import Reservation from './Reservation.jsx'
 
 export default class App extends Component {
 
   render() {
     return (
-      <div className="container">
-        <DatePickerForm />
+      <div>
+        <Reservation />
       </div>
     )
   }
